refactor: usa doubleValue() en cuotaCongelada para compatibilidad con BigDecimal (prep merge Daniel)

+++ src/main/java/com/bolirana/backend/service/ApuestaService.java
@@ -80,7 +80,7 @@ public class ApuestaService {
 
         apuesta.setApostador(apostador);
         apuesta.setOpcion(opcion);
-        apuesta.setCuotaCongelada(opcion.getCuotaActual());
+        apuesta.setCuotaCongelada(opcion.getCuotaActual().doubleValue());
         apuesta.setEstado(EstadoApuesta.REGISTRADA);
 
         Apuesta apuestaGuardada = apuestaRepository.save(apuesta);
